WOOOHOOOO
Finally everything is working as intended.

open websocket -> send live audio binaries to backend -> backend sends to aws -> aws generates transcripts and the final string is sent to the client and closes the websocket connection

the final transcript is received by the client and triggers a POST req for the LLM analyzation using json.
the json object is sent back and rendered on the DOM.

possible update for the future:

Have live transcripts visible to the client
update DOMs